requirements.txt
numpy
pillow
tqdm
matplotlib
opencv-python
tensorboard
tensorboardX